+++ forge/src/main/java/me/codexadrian/tempad/common/utils/forge/TeleportUtilsImpl.java
@@ -1,0 +1,9 @@
+package me.codexadrian.tempad.common.utils.forge;
+
+import net.minecraftforge.fml.ModList;
+
+public class TeleportUtilsImpl {
+    public static boolean isBaubleModLoaded() {
+        return ModList.get().isLoaded("curios");
+    }
+}