+++ changelog.md
@@ -1,3 +1,4 @@
 # 2.3.2
 
 - Add null check to Fabric Waystones
+- Fix crash when Curios or Trinkets is not present